third_party/blink/web_tests/fast/dom/partial-layout-non-overlay-scrollbars.html

<!DOCTYPE HTML>
<html>
<script>
// Note: We force non-overlay scrollbars with CSS.
if (window.internals)
    internals.useMockOverlayScrollbars();
</script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel="stylesheet" href="../../resources/testharness.css">
<head>
<style>
::-webkit-scrollbar {
    -webkit-appearance: none;
    width: 8px;
}

* {
    margin: 0;
    padding: 0;
}

#makespace {
    height: 5000px;
}
</style>
</head>
<body>
<div id="makespace"></div>
<div id="measure"></div>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();

    var test = async_test("Test that partial layout works with non-overlay scrollbars.");
    setTimeout(function() {
        test.step(function() {
            var measure = document.getElementById('measure');
            var measureWidth = measure.offsetWidth; // Partial layout occurs here!

            assert_equals(measureWidth, document.body.offsetWidth);
            assert_equals(measureWidth, window.innerWidth - 8);
        });

        test.done();
    }, 0);
</script>
</body>
</html>
